+++ lib/yaz0.c
@@ -14,7 +14,7 @@ zelda64_yaz0_header_t zelda64_get_yaz0_header(const uint8_t *buf, size_t size) {
     return header;
 }
 
-void zelda64_yaz0_decompress(uint8_t *dest, uint64_t dest_size, const uint8_t *src) {
+void zelda64_yaz0_decompress(uint8_t *dest, size_t dest_size, const uint8_t *src) {
     uint64_t src_index = 16; // skip header
     uint64_t dest_index = 0;
     uint8_t count = 0;
